Fix all remaining mypy errors across backend and opc_service
- Add type guards, explicit annotations, and narrow optional unions
- Fix sort keys, Path/NodeId constructors, and chromadb stub types
- Resolve variable shadowing in opc write route and auditor flow
- Restore StateInput typing in nodes, Callable annotations, quant overrides

+++ backend/company/budget.py
@@ -47,7 +47,7 @@ class CostTracker:
         image_tokens: int = 0,
         audio_tokens: int = 0,
         embedding_tokens: int = 0,
-        **extra: float,
+        **extra: int,
     ) -> float:
         """根據各收費項用量估算成本（USD）。"""
         return estimate_usage_cost(
